pyproject.toml
[project]
name = "rent-scooter"
version = "0.1.0"
description = "Flask Rent Scooter Management System"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "flask-migrate>=4.0.0",
    "pymysql>=1.1.0",
    "cryptography>=42.0.0",
    "pillow>=10.0.0",
    "flask-login>=0.6.3",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
